Move package cliconfig to cli/config
I felt it made more sence 👼
