+++ cli/config/credentials/default_store_windows.go
